Refactor financial transactions to use decimal type
- Updated appsettings.json to change NbpSyncCronExpression.
- Introduced NewTransactionAddCommand and handler for adding transactions.
- Added NewTransactionConvert command and handler for currency conversion.
- Changed Value property type from double to decimal in multiple files for precision.
- Updated validation logic to accommodate new decimal types.
- Created new endpoints for currency conversion and updated related documentation.
- Ensured consistency across Infrastructure and Domain layers for financial calculations.

+++ Application/CQRS/Transaction/Commands/NewTransactionSubtract/NewTransactionSubtractInput.cs
@@ -4,5 +4,5 @@ public record NewTransactionSubtractInput
 (
     string WalletId,
     string Code,
-    double Value
+    decimal Value
 );
